Switch to using tmp dirs; add description and thumnail options
Downloads initial files and fragments to a tmp directory created inside the
final directory. The tmp directory uses the video ID as a prefix. This
replaces the old static audio and video download directories.

Added --write-description and --write-thumbnail options. These are analogous
to youtube-dl's options; writing the video description to a .description file
and the thumbnail to a .jpeg file, respectively.